+++ project/app/components/othello/othello_graphics.cpp
@@ -8,7 +8,8 @@ namespace fc
         const int y = grid_size * (row + 1);
         static constexpr uint32_t black = 0x000000ff;
         static constexpr uint32_t white = 0xd0d0d0ff;
-        image.draw_circle(x, y, disk_radius, is_black ? black : white);
+        image.draw_circle(static_cast<float>(x), static_cast<float>(y),
+            disk_radius, is_black ? black : white);
     }
 
     void OthelloGraphics::draw_hint(gfx::Image& image, const int row, const int column, const bool is_previous)
@@ -17,7 +18,8 @@ namespace fc
         const int y = grid_size * (row + 1);
         static constexpr uint32_t red = 0x7f0000ff;
         static constexpr uint32_t green = 0x007f20ff;
-        image.draw_circle(x, y, hint_radius, is_previous ? red : green);
+        image.draw_circle(static_cast<float>(x), static_cast<float>(y),
+            hint_radius, is_previous ? red : green);
     }
 
     void OthelloGraphics::save_board(const fs::path& path, const State& state,
